backend/requirements.txt
python-dotenv
langgraph
pandas
langchain-core
langchain-openai
uvicorn
fastapi
python-multipart
tidb-vector
sqlalchemy
pymysql
langchain_community
sqlmodel